requirements.txt
python-dateutil>=2.8.2
pandas==1.3.4
numpy>=1.21
scipy>=1.7.1
matplotlib>=3.4
pyyaml>=6.0
tqdm
